Go back to brown color instead of green and white
