+++ app/src/main/res/layout/fragment_list_view.xml
@@ -86,13 +86,9 @@
                 android:choiceMode="singleChoice"
                 android:divider="@color/divider"
                 android:listSelector="#cfcfcf"
+                android:background="#f5f1e4"
                 android:id="@+id/list_view"
                 android:visibility="gone" />
-                <!-- android:divider="@android:color/transparent"
-                android:dividerHeight="0dp"
-                android:background="#f5f1e4"
-                android:listSelector="#cfcfcf" -->
-
 
     </android.support.v4.widget.SwipeRefreshLayout>
