+++ app/src/main/res/layout/fragment_preferences.xml
@@ -22,8 +22,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@android:id/list"
-    android:background="#FFFFFF"
     android:divider="@color/divider"
     android:dividerHeight="1dp"
+    android:background="#f5f1e4"
     android:listSelector="#dfdacf"
     />
