+++ app/src/main/res/menu/main.xml
@@ -19,4 +19,29 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".MainActivity" >
+
+    <item android:title="@string/menu"
+        android:icon="@drawable/ic_dots_vertical_white"
+        android:orderInCategory="120"
+        android:showAsAction="always">
+        <menu>
+            <item android:id="@+id/action_restart"
+                android:title="@string/title_restart"
+                android:orderInCategory="110"
+                android:showAsAction="never"
+                android:icon="@drawable/ic_action_replay" />
+
+            <item android:id="@+id/action_power"
+                android:title="@string/title_shutdown"
+                android:orderInCategory="120"
+                android:showAsAction="never"
+                android:icon="@drawable/ic_action_warning" />
+
+            <item android:id="@+id/action_about"
+                android:title="@string/about"
+                android:orderInCategory="130"
+                android:showAsAction="never"
+                android:icon="@drawable/ic_action_about" />
+        </menu>
+    </item>
 </menu>
