+++ app/src/main/res/xml/preferences.xml
@@ -77,7 +77,7 @@
         android:key="emphasize_show_name"
         android:summary="Emphasize the show name in the 'Coming Episodes' view, instead of the episode name."
         android:title="Emphasize Show Name" />
-
+<!--
     <Preference
         android:key="shutdown"
         android:title="Shutdown" />
@@ -89,4 +89,5 @@
     <Preference
         android:key="about"
         android:title="About" />
+        -->
 </PreferenceScreen>
